app/requirements.txt
streamlit
firebase-admin
torch
torchvision
timm
pydicom
fpdf2
Pillow
numpy
matplotlib
opencv-python
grad-cam
scikit-image
python-dotenv
segmentation-models-pytorch
sendgrid
